Move persistence stuff to own file.
